refactor(secondary-screen): extract pure background builders to top-level functions
Move the 7 pure background/artwork builders out of `_SecondaryScreenState`
into `secondary_screen/background_builders.dart` as top-level functions:
`buildBackgroundBytes`, `buildBackground`, `buildDefaultBackground`,
`buildFanartWithLogo`, `buildUnifiedAppBackground`, `buildSystemBackground`,
`buildShaderFallback`.

Each is a pure mapping from a `SecondaryDisplayStateData` snapshot (plus
constant styling) to a widget subtree — no widget-state reads — so the move is
byte-identical bodies with only the private→public rename. All host call sites
(the `ValueListenableBuilder` in `build`, plus `_buildDefaultStaticUI`) now call
the top-level functions. Two now-unused host imports removed
(shader_gif_widget, image_utils). Host 2492 -> 2332 lines.

First of the Phase-2 secondary_screen panel extractions; behaviour unchanged.
